Transformed from Api style to the appropriate style! SECOND TRY
